requirements.txt
crewai==0.86.0
load_dotenv==0.1.0
crewai_tools==0.17.0